video_analytics_bot/requirements.txt
aiogram==2.25.1
aiohttp==3.9.1
asyncpg==0.29.0
python-dotenv==1.0.0
pydantic==2.5.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
openai==1.6.1
langchain==0.1.0
langchain-openai==0.0.2
requests==2.31.0
pytz==2023.3
python-dateutil==2.8.2